% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{calculate_es}
\alias{calculate_es}
\title{Calculate the effect size of a query}
\usage{
calculate_es(x, ...)
}
\arguments{
\item{x}{an S3 object return from a query, either by the function \code{\link[=query]{query()}} or underlying functions such as \code{\link[=mac]{mac()}}}

\item{...}{additional parameters for the effect size functions}
}
\value{
the effect size
}
\description{
This function calculates the effect of a query.
}
\seealso{
\code{\link[=weat_es]{weat_es()}}, \code{\link[=mac_es]{mac_es()}}, \code{\link[=rnd_es]{rnd_es()}}, \code{\link[=rnsb_es]{rnsb_es()}}
}
\author{
Chung-hong Chan
}
