% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ect.R
\name{ect_es}
\alias{ect_es}
\title{Calculate the Spearman Coefficient of an ECT result}
\usage{
ect_es(x)
}
\arguments{
\item{x}{an ect object from the \code{\link[=ect]{ect()}} function.}
}
\value{
Spearman Coefficient
}
\description{
This functions calculates the Spearman Coefficient of an Embedding Coherence Test. The value ranges from -1 to +1 and a larger value indicates less bias.
}
\references{
Dev, S., & Phillips, J. (2019, April). Attenuating bias in word vectors. In The 22nd International Conference on Artificial Intelligence and Statistics (pp. 879-887). PMLR.
}
\author{
Chung-hong Chan
}
