% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweater.R
\docType{data}
\name{glove_math}
\alias{glove_math}
\title{A subset of the pretrained GLoVE word vectors}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 32 rows and 300 columns.
}
\usage{
glove_math
}
\description{
This is a subset of the original pretrained GLoVE word vectors provided by Pennington et al (2017). The same word vectors were used in Caliskan et al. (2017) to study biases.
}
\references{
Pennington, J., Socher, R., & Manning, C. D. (2014, October). Glove: Global vectors for word representation. In Proceedings of the 2014 conference on empirical methods in natural language processing (EMNLP) (pp. 1532-1543).
Caliskan, A., Bryson, J. J., & Narayanan, A. (2017). Semantics derived automatically from language corpora contain human-like biases. Science, 356(6334), 183-186.
}
\keyword{datasets}
