% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\docType{data}
\name{googlenews}
\alias{googlenews}
\title{A subset of the pretrained word2vec word vectors}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 116 rows and 300 columns.
}
\usage{
googlenews
}
\description{
This is a subset of the original pretrained word2vec word vectors trained on Google News. The same word vectors were used in Garg et al. (2018) to study biases.
}
\references{
Garg, N., Schiebinger, L., Jurafsky, D., & Zou, J. (2018). Word embeddings quantify 100 years of gender and ethnic stereotypes. Proceedings of the National Academy of Sciences, 115(16), E3635-E3644. \doi{10.1073/pnas.1720347115}
}
\keyword{datasets}
