% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nas.R
\name{nas}
\alias{nas}
\title{Calculate Normalized Association Score}
\usage{
nas(w, S, A, B, verbose = FALSE)
}
\arguments{
\item{w}{a numeric matrix of word embeddings (e.g. from rsparse::GloVe)}

\item{S}{a character vector of the first set of target words. In an example of studying gender stereotype, it can include occupations such as programmer, engineer, scientists...}

\item{A}{a character vector of the first set of attribute words. In an example of studying gender stereotype, it can include words such as man, male, he, his.}

\item{B}{a character vector of the second set of attribute words. In an example of studying gender stereotype, it can include words such as woman, female, she, her.}

\item{verbose}{logical, whether to display information}
}
\value{
A list with class \code{"nas"} containing the following components:
\describe{
\item{\code{$P}}{a vector of normalized association score for every word in S}
\item{\code{$raw}}{a list of raw results used for calculating normalized association scores}
\item{\code{$S}}{the input S}
\item{\code{$A}}{the input A}
\item{\code{$B}}{the input B}
}
}
\description{
This functions quantifies the bias in a set of word embeddings by Caliskan et al (2017). In comparison to WEAT introduced in the same paper, this method is more suitable for continuous ground truth data. See Figure 1 and Figure 2 of the original paper.
}
