% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R, R/s3.R
\name{plot_bias}
\alias{plot_bias}
\alias{plot.sweater}
\title{Visualize the bias of words in S}
\usage{
plot_bias(x)

\method{plot}{sweater}(x, ...)
}
\arguments{
\item{x}{an S3 object returned from mac, rnd, semaxis, nas or rnsb}

\item{...}{other parameters}
}
\value{
a plot
}
\description{
For \code{ect}, this function calls \code{\link[=plot_ect]{plot_ect()}}. For \code{mac}, \code{rnd}, and \code{semaxis}, this function plots the bias of words in \code{S} as a Cleveland Dot Plot.
}
\author{
Chung-hong Chan
}
