% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ect.R
\name{plot_ect}
\alias{plot_ect}
\title{Plot an ECT result on a two-dimensional plane}
\usage{
plot_ect(x, ...)
}
\arguments{
\item{x}{an ect object from the \link{ect} function.}

\item{...}{additional parameters to the underlying \code{\link[=plot]{plot()}} function}
}
\value{
a plot
}
\description{
This functions plot the words in \code{S_words} on a 2D plane according to their association with the average vectors of \code{A_words} and \code{B_words}. A equality line is also added. Words along the equality line have less bias. Words located on the upper side of the equality line have a stronger association with \code{A_words} and vice versa.
}
\author{
Chung-hong Chan
}
