% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{read_word2vec}
\alias{read_word2vec}
\title{A helper function for reading word2vec format}
\usage{
read_word2vec(x)
}
\arguments{
\item{x}{path to your text file}
}
\value{
a dense matrix
}
\description{
This function reads word2vec text format and return a dense matrix that can be used by this package.
The file can have or have not the "verification line", i.e. the first line contains the dimensionality of the matrix. If the verification line exists, the function will check the returned matrix for correctness.
}
\author{
Chung-hong Chan
}
