% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnsb.R
\name{rnsb}
\alias{rnsb}
\title{Relative Negative Sentiment Bias}
\usage{
rnsb(w, S, A, B, levels = 1)
}
\arguments{
\item{w}{a numeric matrix of word embeddings (e.g. from rsparse::GloVe)}

\item{S}{a character vector of a set of target words. In an example of studying gender stereotype, it can include occupations such as programmer, engineer, scientists.}

\item{A}{a character vector of the first set of attribute words. In an example of studying gender stereotype, it can include words such as man, male, he, his.}

\item{B}{a character vector of the second set of attribute words. In an example of studying gender stereotype, it can include words such as woman, female, she, her.}

\item{levels}{an integer vector: if S is a quanteda dictionary, this value indicates the levels of analysis.}
}
\value{
A list with class \code{"rnsb"} containing the following components:
\describe{
\item{\code{$classifer}}{ a logistic regression model with L2 regularization trained with LiblineaR}
\item{\code{$A}}{the input A}
\item{\code{$B}}{the input B}
\item{\code{$S}}{the input S}
\item{\code{$P}}{the predicted negative sentiment probabilities}
}
\code{\link{rnsb_es}} can be used to obtain the effect size of the test.
}
\description{
This function estimate the Relative Negative Sentiment Bias (RNSB) of word embeddings (Sweeney & Najafian, 2019).
}
\examples{
S <- c("swedish", "irish", "mexican", "chinese", "filipino", "german", "english", "french", "norwegian", "american", "indian", "dutch", "russian", "scottish", "italian")
sn <- rnsb(w = glove_sweeney, S = S, A = bing_pos, B = bing_neg)
sort(sn$P)
}
\references{
Sweeney, C., & Najafian, M. (2019, July). A transparent framework for evaluating unintended demographic bias in word embeddings. In Proceedings of the 57th Annual Meeting of the Association for Computational Linguistics (pp. 1662-1667).
}
