% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semaxis.R
\docType{data}
\name{small_reddit}
\alias{small_reddit}
\title{A subset of the pretrained word2vec word vectors on Reddit}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 106 rows and 300 columns.
}
\usage{
small_reddit
}
\description{
This is a subset of the pretrained word2vec word vectors on Reddit provided by An et al. (2018). With this dataset, you can try with the "l" parameter of \code{\link[=semaxis]{semaxis()}} up to 10.
}
\references{
An, J., Kwak, H., & Ahn, Y. Y. (2018). \href{https://arxiv.org/abs/1806.05521}{SemAxis: A lightweight framework to characterize domain-specific word semantics beyond sentiment.} arXiv preprint arXiv:1806.05521.
}
\keyword{datasets}
