% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweater.R
\name{sweater_es}
\alias{sweater_es}
\title{Calculation of effect size}
\usage{
sweater_es(sweater_obj, standardize = TRUE)
}
\arguments{
\item{standardize}{a boolean to denote division of the difference by the standard division. The standardized version can be interpreted the same way as Cohen's d.}
}
\description{
Calculation of effect size
}
