% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac.R
\name{mac_es}
\alias{mac_es}
\title{Calculation of MAC}
\usage{
mac_es(x)
}
\arguments{
\item{x}{an object from the function \link{mac}}
}
\value{
Mean of all cosine similarity values
}
\description{
This function calculates the mean of cosine distance values
}
