% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{plot_bias}
\alias{plot_bias}
\title{Visualize the bias of words in S}
\usage{
plot_bias(x)
}
\arguments{
\item{x}{an S3 object returned from mac, rnd, semaxis, nas or rnsb}
}
\value{
a plot
}
\description{
This function plots the bias of words in \code{S} as a Cleveland Dot Plot.
}
\author{
Chung-hong Chan
}
