% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd.R
\name{rnd_es}
\alias{rnd_es}
\title{Calculation of sum of all relative norm distances}
\usage{
rnd_es(x)
}
\arguments{
\item{x}{an object from the function \link{rnd}}
}
\value{
Sum of all relative norm distances
}
\description{
This function calculates the sum of all relative norm distances from the relative norm distance test.
}
