% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnd.R
\name{rnd_es}
\alias{rnd_es}
\title{Calculation of sum of all relative norm distances}
\usage{
rnd_es(x)
}
\arguments{
\item{x}{an object from the function \link{rnd}}
}
\value{
Sum of all relative norm distances
}
\description{
This function calculates the sum of all relative norm distances from the relative norm distance test. If possible, please use \code{\link[=calculate_es]{calculate_es()}} instead.
}
\references{
Garg, N., Schiebinger, L., Jurafsky, D., & Zou, J. (2018). Word embeddings quantify 100 years of gender and ethnic stereotypes. Proceedings of the National Academy of Sciences, 115(16), E3635-E3644. \doi{10.1073/pnas.1720347115}
}
