% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sweater.R
\name{weat_es}
\alias{weat_es}
\title{Calculation of effect size}
\usage{
weat_es(sweater_obj, standardize = TRUE, r = FALSE)
}
\arguments{
\item{standardize}{a boolean to denote whether to correct the difference by the standard division. The standardized version can be interpreted the same way as Cohen's d.}

\item{r}{a boolean to denote whether convert the effect size to Pearson's correlation coefficient.}
}
\description{
This function calculates the effect size from a sweater object. The original implementation in Caliskan et al. (2017) assumes the numbers of words in S and in T must be equal. The current implementation eases this assumption by adjusting the variance with the difference in sample sizes. It is also possible to convert the Cohen's d to Pearson's correlation coefficient (r).
}
